+++ javaparser-core/src/main/java/com/github/javaparser/ast/body/FieldDeclaration.java
@@ -30,6 +30,7 @@ import java.util.List;
 
 import com.github.javaparser.ASTHelper;
 import com.github.javaparser.Range;
+import com.github.javaparser.ast.Modifier;
 import com.github.javaparser.ast.comments.JavadocComment;
 import com.github.javaparser.ast.expr.AnnotationExpr;
 import com.github.javaparser.ast.nodeTypes.NodeWithJavaDoc;
